% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/games.R
\name{get_boxscore}
\alias{get_boxscore}
\title{Get box score for specified game ID}
\usage{
get_boxscore(gameID = NULL)
}
\arguments{
\item{gameID}{A ten digit game ID. The first 4 digits identify the season of the game (ie. 2017 for the 2017-2018 season). The next 2 digits give the type of game, where 01 = preseason, 02 = regular season, 03 = playoffs, 04 = all-star. The final 4 digits identify the specific game number. For regular season and preseason games, this ranges from 0001 to the number of games played. (1271 for seasons with 31 teams (2017 and onwards) and 1230 for seasons with 30 teams). For playoff games, the 2nd digit of the specific number gives the round of the playoffs, the 3rd digit specifies the matchup, and the 4th digit specifies the game (out of 7).}
}
\value{
The box score for the specified \code{gameID}. Returned as a list.
}
\description{
Get box score for specified game ID
}
\examples{
get_boxscore(gameID = 2017030111)
}
